Testcase for Date / Time Manipulation via HSQLDB added
  - added entities to ao-modules entities
  - Renamed getPause to getPauseMinutes (causes HSQLDB error)

+++ src/main/java/org/catrobat/confluence/services/TimesheetEntryService.java
@@ -14,5 +14,5 @@ import org.catrobat.confluence.activeobjects.TimesheetEntry;
 public interface TimesheetEntryService
 {
 	TimesheetEntry add(Timesheet sheet, Date begin, Date end, Category category,
-					String description, long pause, Project project);
+					String description, int pause, Project project);
 }
